Switch to GitHub Actions
https://github.com/open-contracting/cove-oc4ids/issues/43

This also removes tests/test_browser.py.
This was having problems on GitHub, and was introdued by
https://github.com/open-contracting/cove-oc4ids/commit/7a73219e1e3af156367186a72058ec210e57d975
It's purpose was to simply make sure the Selenium tests run.
But we have test_footer_oc4ids in cove_oc4ids/test_functional.py
And that will check the server anyway, so it's ok.